Added feature control by loggedin status: Edit, Delete, Create. Create send userId of logged in user just before form submit

+++ src/components/streams/StreamList.js
@@ -9,21 +9,35 @@ class StreamList extends React.Component {
 		this.props.getStreams();
 	}
 
+	renderEditButtons = (stream) => {
+		let { loginStatus } = this.props;
+
+		if (
+			loginStatus &&
+			loginStatus.isLoggedIn &&
+			stream.userId === loginStatus.userId
+		) {
+			return (
+				<div className="right floated content">
+					<div className="ui button">Edit</div>
+					<div
+						className="ui button"
+						onClick={() => {
+							this.props.deleteStream(stream.id);
+						}}
+					>
+						Delete
+					</div>
+				</div>
+			);
+		}
+		return null;
+	};
 	renderStreams = function (streams) {
 		return streams.map((stream) => {
 			return (
 				<div className="ui item padded segment" key={stream.id}>
-					<div className="right floated content">
-						<div className="ui button">Edit</div>
-						<div
-							className="ui button"
-							onClick={() => {
-								this.props.deleteStream(stream.id);
-							}}
-						>
-							Delete
-						</div>
-					</div>
+					{this.renderEditButtons(stream)}
 					<div className="content">
 						<div className="header">{stream.title}</div>
 						<div className="description">{stream.description}</div>
@@ -33,16 +47,27 @@ class StreamList extends React.Component {
 		});
 	};
 
+	renderCreateIfLoggedIn = () => {
+		let { loginStatus } = this.props;
+		console.log("STATUS:", loginStatus);
+		if (loginStatus && loginStatus.isLoggedIn === true) {
+			return (
+				<Link className="" to="/streams/new">
+					<div className="ui right floated button green">
+						Add Stream
+					</div>
+				</Link>
+			);
+		}
+		return null;
+	};
+
 	render() {
 		return (
 			<div>
 				<div className="ui clearing green padded segment">
 					<h1>Streams</h1>
-					<Link className="" to="/streams/new">
-						<div className="ui right floated button green">
-							Add Stream
-						</div>
-					</Link>
+					{this.renderCreateIfLoggedIn()}
 				</div>
 				<div className="ui list relaxed middle aligned">
 					{this.renderStreams(this.props.streams)}
@@ -53,7 +78,11 @@ class StreamList extends React.Component {
 }
 
 const mapStateToProps = function (state, oldProps) {
-	return { streams: Object.values(state.streams) };
+	console.log("STATE:", state);
+	return {
+		streams: Object.values(state.streams),
+		loginStatus: state.loginStatus,
+	};
 };
 
 const reduxStreamlist = connect(mapStateToProps, { getStreams, deleteStream })(
